Only displays presets now.
Updated to v1.1.0 with a slightly different app flow. Now instead of displaying "PRESET #X" when there are unset presets, the app simply passes over that preset slot. Also displays a "No Presets Set" display when there are no available presets. Additionally, added an error screen for when the application detects that persistant storage is broken on the watch.